productivity-agent-system/requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
streamlit==1.29.0
plotly==5.18.0
pydantic==2.5.3

# Optional: Only needed if using cloud providers instead of Ollama
# Uncomment if using Groq
# langchain-groq==0.0.1

# Uncomment if using Google Gemini
# langchain-google-genai==0.0.5